fix: apply @KaterinaWire's changes
