+++ Wire-iOS/Sources/Components/ShareViewController/ShareViewController+Views.swift
@@ -18,7 +18,6 @@
 
 import Foundation
 import UIKit
-import WireCommonComponents
 
 extension ShareViewController {
