+++ Wire-iOS/Sources/Helpers/UIBarButtonItem+StyleKit.swift
@@ -39,7 +39,7 @@ extension UIBarButtonItem {
 
     static func createUpdatedCloseItem() -> UIBarButtonItem {
         let item = UIBarButtonItem(icon: .cross, target: nil, action: nil)
-        item.tintColor = SemanticColors.IconColors.foregroundCellIconActive
+        item.tintColor = SemanticColors.Icon.foregroundCellIconActive
         item.accessibilityIdentifier = "close"
         item.accessibilityLabel = "general.close".localized
         return item
