+++ Wire-iOS/Sources/UserInterface/AddContacts/AddParticipantsViewModel.swift
@@ -80,7 +80,7 @@ struct AddParticipantsViewModel {
         switch context {
         case .add:
             let item = UIBarButtonItem(icon: .cross, target: target, action: action)
-            item.tintColor = SemanticColors.IconColors.foregroundCellIconActive
+            item.tintColor = SemanticColors.Icon.foregroundCellIconActive
             item.accessibilityIdentifier = "close"
             return item
         case .create(let values):
