+++ Wire-iOS/Sources/UserInterface/Components/Views/IconImageView/IconImageView.swift
@@ -43,7 +43,7 @@ extension IconImageStyle {
 
 class IconImageView: UIImageView {
     private(set) var size: StyleKitIcon.Size = .tiny
-    private(set) var color: UIColor = SemanticColors.IconColors.foregroundCellIconActive
+    private(set) var color: UIColor = SemanticColors.Icon.foregroundCellIconActive
     private(set) var style: IconImageStyle?
 
     override init(frame: CGRect) {
