+++ Wire-iOS/Sources/UserInterface/GroupDetails/Cells/GroupDetailsGuestOptionsCell.swift
@@ -41,7 +41,7 @@ final class GroupDetailsGuestOptionsCell: GroupDetailsDisclosureOptionsCell {
 
     override func applyColorScheme(_ colorSchemeVariant: ColorSchemeVariant) {
         super.applyColorScheme(colorSchemeVariant)
-        iconColor = SemanticColors.IconColors.foregroundCellIconActive
+        iconColor = SemanticColors.Icon.foregroundCellIconActive
         guard let iconColor = iconColor else { return }
         icon = StyleKitIcon.guest.makeImage(size: .tiny,
                                             color: iconColor).withRenderingMode(.alwaysTemplate)
@@ -50,8 +50,8 @@ final class GroupDetailsGuestOptionsCell: GroupDetailsDisclosureOptionsCell {
     override var isHighlighted: Bool {
         didSet {
             backgroundColor = isHighlighted
-            ? SemanticColors.BackgroundColors.backgroundUserCellHightLighted
-            : SemanticColors.BackgroundColors.backgroundUserCell
+            ? SemanticColors.View.Background.backgroundUserCellHightLighted
+            : SemanticColors.View.Background.backgroundUserCell
         }
     }
