+++ Wire-iOS/Sources/UserInterface/GroupDetails/Cells/GroupDetailsNotificationOptionsCell.swift
@@ -38,7 +38,7 @@ final class GroupDetailsNotificationOptionsCell: GroupDetailsDisclosureOptionsCe
 
     override func applyColorScheme(_ colorSchemeVariant: ColorSchemeVariant) {
         super.applyColorScheme(colorSchemeVariant)
-        iconColor = SemanticColors.IconColors.foregroundCellIconActive
+        iconColor = SemanticColors.Icon.foregroundCellIconActive
         guard let iconColor = iconColor else { return }
 
         icon = StyleKitIcon.alerts.makeImage(size: .tiny,
@@ -48,8 +48,8 @@ final class GroupDetailsNotificationOptionsCell: GroupDetailsDisclosureOptionsCe
     override var isHighlighted: Bool {
         didSet {
             backgroundColor = isHighlighted
-            ? SemanticColors.BackgroundColors.backgroundUserCellHightLighted
-            : SemanticColors.BackgroundColors.backgroundUserCell
+            ? SemanticColors.View.Background.backgroundUserCellHightLighted
+            : SemanticColors.View.Background.backgroundUserCell
         }
     }
