+++ Wire-iOS/Sources/UserInterface/GroupDetails/Cells/GroupDetailsReceiptOptionsCell.swift
@@ -33,7 +33,7 @@ final class GroupDetailsReceiptOptionsCell: IconToggleCell {
 
     override func applyColorScheme(_ colorSchemeVariant: ColorSchemeVariant) {
         super.applyColorScheme(colorSchemeVariant)
-        iconColor = SemanticColors.IconColors.foregroundCellIconActive
+        iconColor = SemanticColors.Icon.foregroundCellIconActive
         guard let iconColor = iconColor else { return }
         icon = StyleKitIcon.eye.makeImage(
             size: .tiny,
@@ -44,8 +44,8 @@ final class GroupDetailsReceiptOptionsCell: IconToggleCell {
     override var isHighlighted: Bool {
         didSet {
             backgroundColor = isHighlighted
-            ? SemanticColors.BackgroundColors.backgroundUserCellHightLighted
-            : SemanticColors.BackgroundColors.backgroundUserCell
+            ? SemanticColors.View.Background.backgroundUserCellHightLighted
+            : SemanticColors.View.Background.backgroundUserCell
         }
     }
 }
