+++ Wire-iOS/Sources/UserInterface/GroupDetails/GroupDetailsViewController.swift
@@ -96,7 +96,7 @@ final class GroupDetailsViewController: UIViewController, ZMConversationObserver
     override func viewDidLoad() {
         super.viewDidLoad()
         title = "participants.title".localized(uppercased: true)
-        view.backgroundColor = SemanticColors.BackgroundColors.backgroundViewDefault
+        view.backgroundColor = SemanticColors.View.Background.backgroundViewDefault
     }
 
     override func viewWillAppear(_ animated: Bool) {
