+++ Wire-iOS/Sources/UserInterface/StartUI/StartUI/SearchResultsView.swift
@@ -40,7 +40,7 @@ final class SearchResultsView: UIView {
         collectionViewLayout.minimumLineSpacing = 0
 
         collectionView = UICollectionView(frame: CGRect.zero, collectionViewLayout: collectionViewLayout)
-        collectionView.backgroundColor = SemanticColors.BackgroundColors.backgroundViewDefault
+        collectionView.backgroundColor = SemanticColors.View.Background.backgroundViewDefault
         collectionView.allowsMultipleSelection = true
         collectionView.keyboardDismissMode = .onDrag
         collectionView.bounces = true
