+++ Wire-iOS/Sources/UserInterface/UserProfile/Devices/UserClientListViewController.swift
@@ -81,7 +81,7 @@ final class UserClientListViewController: UIViewController, UICollectionViewDele
         collectionView.register(CollectionViewCellAdapter.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: CollectionViewCellAdapter.zm_reuseIdentifier)
 
         view.addSubview(collectionView)
-        view.backgroundColor = SemanticColors.BackgroundColors.backgroundViewDefault
+        view.backgroundColor = SemanticColors.View.Background.backgroundViewDefault
     }
 
     private func createConstraints() {
